+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardFooter/planDisplayCardFooter.types.ts
@@ -1,3 +1,9 @@
-import { PropsWithChildren } from "react";
+import { PropsWithChildren, ReactNode } from "react";
 
-export type PlanDisplayCardFooterProperties = PropsWithChildren;
+export type PlanDisplayCardFooterProperties = PropsWithChildren<{
+  /**
+   * Icon rendered before the footer's content.
+   * @TJS-type React.ReactNode
+   */
+  icon?: ReactNode;
+}>;