+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayPrice/index.ts
@@ -1,0 +1,5 @@
+import { PlanDisplayPrice } from "./PlanDisplayPrice";
+
+export { PlanDisplayPrice };
+
+export default PlanDisplayPrice;